refactor: 테스트코드 클래스별로 분리
